Split out S2Metric
